Isolated qliu

+++ util/Radiance_Monitor/image_gen/ush/mk_angle_plots.sh
@@ -12,6 +12,7 @@
 
 set -ax
 date
+#export list=$listvar
 
 export NUM_CYCLES=${NUM_CYCLES:-121}
 
@@ -39,8 +40,10 @@ for type in ${SATYPE}; do
    done=0
    test_day=$PDATE
    ctr=$ndays
+#   echo "before while loop, found, done = $found, $done"
 
    while [[ $found -eq 0 && $done -ne 1 ]]; do
+#      echo "top of while loop"
 
       pdy=`echo $test_day|cut -c1-8`    
       if [[ -s ${TANKDIR}/radmon.${pdy}/angle.${type}.ctl.${Z} ]]; then
@@ -64,6 +67,22 @@ for type in ${SATYPE}; do
    if [[ -s ${imgndir}/${type}.ctl.${Z} || -s ${imgndir}/${type}.ctl ]]; then
       allmissing=0
       found=1
+
+#   elif [[ -s ${TANKDIR}/radmon.${PDY}/angle.${type}.ctl || -s ${TANKDIR}/radmon.${PDY}/angle.${type}.ctl.${Z} ]]; then
+#      $NCP ${TANKDIR}/radmon.${PDY}/angle.${type}.ctl.${Z} ${imgndir}/${type}.ctl.${Z}
+#      if [[ ! -s ${imgndir}/${type}.ctl.${Z} ]]; then
+#         $NCP ${TANKDIR}/radmon.${PDY}/angle.${type}.ctl ${imgndir}/${type}.ctl
+#      fi
+#      allmissing=0
+#      found=1
+#
+#   elif [[ -s ${tankdir}/${type}.ctl.${Z} || -s ${tankdir}/${type}.ctl  ]]; then
+#      $NCP ${tankdir}/${type}.ctl* ${imgndir}/.
+#      allmissing=0
+#      found=1
+#
+#   else
+#      echo WARNING:  unable to locate ${type}.ctl
    fi
 done
 
@@ -72,17 +91,23 @@ if [[ $allmissing = 1 ]]; then
    exit
 fi
 
+# TESTING
+#export SATYPE="iasi_metop-a sndrd1_g15 sndrd2_g15"
+#export SATYPE="sndrd1_g15"
+
 #-------------------------------------------------------------------
 #   Update the time definition (tdef) line in the angle control 
 #   files. Conditionally rm "cray_32bit_ieee" from the options line.
  
+#thirtydays=`$NDATE -720 $PDATE`
+
 for type in ${SATYPE}; do
    if [[ -s ${imgndir}/${type}.ctl.${Z} ]]; then
      ${UNCOMPRESS} ${imgndir}/${type}.ctl.${Z}
    fi
-   ${IG_SCRIPTS}/update_ctl_tdef.sh ${imgndir}/${type}.ctl ${START_DATE} ${NUM_CYCLES}
+   ${SCRIPTS}/update_ctl_tdef.sh ${imgndir}/${type}.ctl ${START_DATE} ${NUM_CYCLES}
 
-   if [[ $MY_MACHINE = "wcoss" || $MY_MACHINE = "zeus" || $MY_MACHINE = "theia" ]]; then
+   if [[ $MY_MACHINE = "wcoss" ]]; then
       sed -e 's/cray_32bit_ieee/ /' ${imgndir}/${type}.ctl > tmp_${type}.ctl
       mv -f tmp_${type}.ctl ${imgndir}/${type}.ctl
    fi
@@ -105,7 +130,7 @@ ${COMPRESS} -f ${imgndir}/*.ctl
 #-------------------------------------------------------------------
 #   Rename PLOT_WORK_DIR to angle subdir.
 #
-export PLOT_WORK_DIR="${PLOT_WORK_DIR}/plotangle_${RADMON_SUFFIX}"
+export PLOT_WORK_DIR="${PLOT_WORK_DIR}/plotangle_${SUFFIX}"
 
 if [[ -d $PLOT_WORK_DIR ]]; then
    rm -f $PLOT_WORK_DIR
@@ -117,58 +142,52 @@ cd $PLOT_WORK_DIR
   #-----------------------------------------------------------------
   # Loop over satellite types.  Submit job to make plots.
   #
+export listvar=RAD_AREA,LOADLQ,PDATE,START_DATE,NUM_CYCLES,NDATE,TANKDIR,IMGNDIR,PLOT_WORK_DIR,EXEDIR,LOGDIR,SCRIPTS,GSCRIPTS,STNMAP,GRADS,GADDIR,USER,STMP_USER,PTMP_USER,USER_CLASS,SUB,SUFFIX,SATYPE,NCP,Z,COMPRESS,UNCOMPRESS,PLOT_ALL_REGIONS,SUB_AVG,listvar
 
-list="count penalty omgnbc total omgbc fixang lapse lapse2 const scangl clw cos sin emiss ordang4 ordang3 ordang2 ordang1"
+list="count penalty omgnbc total omgbc fixang lapse lapse2 const scangl clw"
 
-  if [[ ${MY_MACHINE} = "wcoss" || ${MY_MACHINE} = "cray" ]]; then
+  if [[ ${MY_MACHINE} = "ccs" || ${MY_MACHINE} = "wcoss" ]]; then
      suffix=a
      cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}
-     jobname=plot_${RADMON_SUFFIX}_ang_${suffix}
-     logfile=$LOGdir/plot_angle_${suffix}.log
+     jobname=plot_${SUFFIX}_ang_${suffix}
+     logfile=$LOGDIR/plot_angle_${suffix}.log
 
      rm -f $cmdfile
      rm -f $logfile
 
-     rm $LOGdir/plot_angle_${suffix}.log
-
+     rm $LOGDIR/plot_angle_${suffix}.log
+#>$cmdfile
      for type in ${SATLIST}; do
-       echo "$IG_SCRIPTS/plot_angle.sh $type $suffix '$list'" >> $cmdfile
+       echo "$SCRIPTS/plot_angle.sh $type $suffix '$list'" >> $cmdfile
      done
      chmod 755 $cmdfile
-     echo "CMDFILE:  $cmdfile"
 
      ntasks=`cat $cmdfile|wc -l `
 
-     if [[ $PLOT_ALL_REGIONS -eq 1 || $ndays -gt 30 ]]; then
-        wall_tm="2:30"
+     if [[ $MY_MACHINE = "wcoss" ]]; then
+        $SUB -q dev -o ${logfile} -W 1:45 -R affinity[core] -J ${jobname} $cmdfile
      else
-        wall_tm="1:45"
+        $SUB -a $ACCOUNT -e $listvar -j ${jobname} -u $USER -t 0:45:00 -o ${logfile} -p $ntasks/1/N -q dev -g ${USER_CLASS}  /usr/bin/poe -cmdfile $cmdfile -pgmmodel mpmd -ilevel 2 -labelio yes -stdoutmode ordered
      fi
-
-     if [[ ${MY_MACHINE} = "wcoss" ]]; then
-        $SUB -q $JOB_QUEUE -P $PROJECT -o ${logfile} -M 20000 -W ${wall_tm} -R affinity[core] -J ${jobname} $cmdfile
-     else	# cray
-        $SUB -q $JOB_QUEUE -P $PROJECT -o ${logfile} -M 20000 -W ${wall_tm} -J ${jobname} $cmdfile
-     fi
-  else				# Zeus/theia platform
+  else				# Zeus/linux platform
      for sat in ${SATLIST}; do
         suffix=${sat} 
         cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}
-        jobname=plot_${RADMON_SUFFIX}_ang_${suffix}
-        logfile=${LOGdir}/plot_angle_${suffix}.log
+        jobname=plot_${SUFFIX}_ang_${suffix}
+        logfile=${LOGDIR}/plot_angle_${suffix}.log
 
         rm -f $cmdfile
         rm -f $logfile
 
-        echo "$IG_SCRIPTS/plot_angle.sh $sat $suffix '$list'" >> $cmdfile
+        echo "$SCRIPTS/plot_angle.sh $sat $suffix '$list'" >> $cmdfile
 
-        if [[ $PLOT_ALL_REGIONS -eq 1 || $ndays -gt 30 ]]; then
-           wall_tm="5:00:00"
+        if [[ $PLOT_ALL_REGIONS -eq 0 ]]; then
+           wall_tm="2:00:00"
         else
-           wall_tm="2:30:00"
+           wall_tm="4:00:00"
         fi
 
-        $SUB -A $ACCOUNT -l procs=1,walltime=${wall_tm} -N ${jobname} -V -j oe -o ${logfile} ${cmdfile}
+        $SUB -A $ACCOUNT -l procs=1,walltime=${wall_tm} -N ${jobname} -v $listvar -j oe -o ${logfile} ${cmdfile}
      done
   fi
 
@@ -181,58 +200,49 @@ list="count penalty omgnbc total omgbc fixang lapse lapse2 const scangl clw cos 
   #    job for each is necessary.
   #   
 
-echo "starting $bigSATLIST"
+echo starting $bigSATLIST
 
-set -A list count penalty omgnbc total omgbc fixang lapse lapse2 const scangl clw cos sin emiss ordang4 ordang3 ordang2 ordang1
+set -A list count penalty omgnbc total omgbc fixang lapse lapse2 const scangl clw
 
 for sat in ${bigSATLIST}; do
    echo processing $sat in $bigSATLIST
 
    #
-   #  wcoss submit 4 jobs for each $sat
+   #  CCS submit 4 jobs for each $sat
    #
-   if [[ $MY_MACHINE = "wcoss" || $MY_MACHINE = "cray" ]]; then 	
+   if [[ $MY_MACHINE = "ccs" || $MY_MACHINE = "wcoss" ]]; then 	
       batch=1
       ii=0
 
       suffix="${sat}_${batch}"
       cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}
       rm -f $cmdfile
-      jobname=plot_${RADMON_SUFFIX}_ang_${suffix}
-      logfile=${LOGdir}/plot_angle_${suffix}.log
+      jobname=plot_${SUFFIX}_ang_${suffix}
+      logfile=${LOGDIR}/plot_angle_${suffix}.log
 
       while [[ $ii -le ${#list[@]}-1 ]]; do
 
-         echo "$IG_SCRIPTS/plot_angle.sh $sat $suffix ${list[$ii]}" >> $cmdfile
-         ntasks=`cat $cmdfile|wc -l `
-         chmod 755 $cmdfile
+         echo "$SCRIPTS/plot_angle.sh $sat $suffix ${list[$ii]}" >> $cmdfile
+         (( test=ii+1 ))
+         (( test=test%3 ))
 
-         if [[ $PLOT_ALL_REGIONS -eq 1 || $ndays -gt 30 ]]; then
-            wall_tm="3:00"
-         else
-            wall_tm="1:00"
+         if [[ $test -eq 0 || $ii -eq ${#list[@]}-1 ]]; then
+            ntasks=`cat $cmdfile|wc -l `
+            chmod 755 $cmdfile
+
+            if [[ $MY_MACHINE = "ccs" ]]; then
+               $SUB -a $ACCOUNT -e $listvar -j ${jobname} -u $USER -t 1:00:00 -o ${logfile} -p $ntasks/1/N -q dev -g ${USER_CLASS} /usr/bin/poe -cmdfile $cmdfile -pgmmodel mpmd -ilevel 2 -labelio yes -stdoutmode ordered
+            else
+               $SUB -q dev -o ${logfile} -W 1:00 -R affinity[core] -J ${jobname} $cmdfile
+            fi
+            (( batch=batch+1 ))
+
+            suffix="${sat}_${batch}"
+            cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}
+            rm -f $cmdfile
+            jobname=plot_${SUFFIX}_ang_${suffix}
+            logfile=${LOGDIR}/plot_angle_${suffix}.log
          fi
-
-        
-#         mem="6000"
-#         if [[ $batch -eq 1 ]]; then
-            mem="24000"
-#         fi
-
-         if [[ $MY_MACHINE = "wcoss" ]]; then
-            $SUB -q $JOB_QUEUE -P $PROJECT -o ${logfile} -M ${mem} -W ${wall_tm} -R affinity[core] -J ${jobname} $cmdfile
-         else
-            $SUB -q $JOB_QUEUE -P $PROJECT -o ${logfile} -M ${mem} -W ${wall_tm} -J ${jobname} $cmdfile
-         fi
-
-         (( batch=batch+1 ))
-
-         suffix="${sat}_${batch}"
-         cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}
-         rm -f $cmdfile
-         jobname=plot_${RADMON_SUFFIX}_ang_${suffix}
-         logfile=${LOGdir}/plot_angle_${suffix}.log
-
          (( ii=ii+1 ))
       done
 
@@ -244,18 +254,18 @@ for sat in ${bigSATLIST}; do
       while [[ $ii -le ${#list[@]}-1 ]]; do
          cmdfile=${PLOT_WORK_DIR}/cmdfile_pangle_${suffix}_${list[$ii]}
          rm -f $cmdfile
-         logfile=${LOGdir}/plot_angle_${suffix}_${list[$ii]}.log
-         jobname=plot_${RADMON_SUFFIX}_ang_${suffix}_${list[$ii]}
+         logfile=${LOGDIR}/plot_angle_${suffix}_${list[$ii]}.log
+         jobname=plot_${SUFFIX}_ang_${suffix}_${list[$ii]}
 
-         echo "${IG_SCRIPTS}/plot_angle.sh $sat $suffix ${list[$ii]}" >> $cmdfile
+         echo "${SCRIPTS}/plot_angle.sh $sat $suffix ${list[$ii]}" >> $cmdfile
 
-         if [[ $PLOT_ALL_REGIONS -eq 1 || $ndays -gt 30 ]]; then
-            wall_tm="5:00:00"
+         if [[ $PLOT_ALL_REGIONS -eq 0 ]]; then
+            wall_tm="2:00:00"
          else
-            wall_tm="2:30:00"
+            wall_tm="4:00:00"
          fi
 
-         $SUB -A $ACCOUNT -l procs=1,walltime=${wall_tm} -N ${jobname} -V -j oe -o ${logfile} ${cmdfile}
+         $SUB -A $ACCOUNT -l procs=1,walltime=${wall_tm} -N ${jobname} -v $listvar -j oe -o ${logfile} ${cmdfile}
 
          (( ii=ii+1 ))
       done
